DPAD-995: splitting content overlays into different files
