+++ src/experimental/types.d.ts
@@ -9,3 +9,22 @@ export interface VolumeSliderProps {
 	hover: boolean;
 	color?: string;
 }
+
+export interface ContentPlayerOverlayProps {
+	isScrollPlayer: boolean;
+	setDismissed: (dismissed: boolean) => void;
+}
+
+export interface PrerollPlayerOverlayProps {
+	isScrollPlayer: boolean;
+	setDismissed: (dismissed: boolean) => void;
+}
+
+export interface TimeSliderProps {
+	className?: string;
+	interactive?: boolean;
+	railColor?: string;
+	bufferColor?: string;
+	knobColor?: string;
+	progressColor?: string;
+}
